nonnonononononononononononononononnonononononononnonononononononononono
I should have finished that

+++ src/commands/invite-remover.ts
@@ -6,7 +6,7 @@ import {Message, Client} from 'discord.js';
 export default class InviteRemovalCommand {
   constructor(private client: Client) {
     client.on('message', message => {
-      if(message.content.toLowerCase()) {
+      if(message.content.toLowerCase() && false) {
         message.delete();
         message.reply('Oi you, no posting invites.');
       }
